All vars to int

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        float price = 150;
+        int price = 150;
         int every20 = 20;
-        int bonus = (int) (price / every20);
+        int bonus = price / every20;
         System.out.print("Бонусов начислено: " + bonus);
     }
 }
